# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Safety.uz API
  description: Automatically generated OpenAPI 3.0 specification for Safety.uz.
  version: 1.0.0
servers:
  - url: /
    description: Current server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    SessionAuth:
      type: apiKey
      in: cookie
      name: session
...
paths:
...
  /api/news/admin:
    get:
      tags: [admin]
      summary: Api admin news
      operationId: api_admin_news_get
      security:
        - SessionAuth: []
        - BearerAuth: []
      responses:
        '200':
          description: Successful operation
...